[fix] Conflit d'import config (#12)
